jetson/requirements.txt
opencv-python
numpy
python3-zmq
python-dotenv